+++ tools/Cargo.toml
@@ -1,4 +1,5 @@
 [workspace]
+resolver = "2"
 members = ["archiver", "cli", "shaderbuild"]
 
 [workspace.dependencies]
