+++ tools/archiver/Cargo.toml
@@ -2,10 +2,13 @@
 name = "peridot-archiver"
 version = "0.1.0"
 authors = ["S.Percentage <Syn.Tri.Naga@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 [dependencies]
-peridot-archive = { path = "../../modules/archive" }
+peridot-archive = { path = "../../modules/archive", features = [
+    "async-rt-async-std",
+] }
+async-std = { version = "1.13", features = ["attributes"] }
 clap = "2.32"
 glob = "0.2"
 libc = "0.2"
